update Add data into admin
