+++ Example/cwp/myapp/apps.py
