+++ Example/cwp/myapp/forms.py
@@ -1,0 +1,8 @@
+# myapp/forms.py
+from django import forms
+from .models import MyModel
+
+class MyModelForm(forms.ModelForm):
+    class Meta:
+        model = MyModel
+        fields = ['name', 'age']
