+++ Example/cwp/myapp/models.py
@@ -1,0 +1,7 @@
+# myapp/models.py
+from django.db import models
+
+class MyModel(models.Model):
+    name = models.CharField(max_length=255)
+    age = models.IntegerField()
+
