+++ Example/cwp/myapp/templates/myapp/form.html
@@ -1,0 +1,6 @@
+<!-- myapp/templates/myapp/form.html -->
+<form method="post" action="{% url 'form' %}">
+    {% csrf_token %}
+    {{ form.as_p }}
+    <button type="submit">Submit</button>
+</form>
