+++ Example/cwp/myapp/templates/myapp/success.html
@@ -1,0 +1,2 @@
+<!-- myapp/templates/myapp/success.html -->
+<h1>Form submitted successfully!</h1>
