--- Project1/myproject/myapp/forms.py
@@ -1,8 +1,0 @@
-# myapp/forms.py
-from django import forms
-from .models import MyModel
-
-class MyModelForm(forms.ModelForm):
-    class Meta:
-        model = MyModel
-        fields = ['name', 'email', 'message']
