--- Project1/myproject/myapp/models.py
@@ -1,11 +1,0 @@
-# myapp/models.py
-from django.db import models
-
-class MyModel(models.Model):
-    name = models.CharField(max_length=255)
-    email = models.EmailField()
-    message = models.TextField()
-
-    def __str__(self):
-        return self.name
-
