--- Project1/myproject/myapp/templates/myapp/my_form.html
@@ -1,6 +1,0 @@
-<!-- myapp/templates/myapp/my_form.html -->
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <button type="submit">Submit</button>
-</form>
