--- Project1/myproject/myapp/templates/myapp/success.html
@@ -1,2 +1,0 @@
-<!-- myapp/templates/myapp/success.html -->
-<h2>Form submitted successfully!</h2>
